# gc_registry/static/openapi.yaml
components:
  schemas:
    Account:
      properties:
        id:
          format: uuid
          title: Account Id
          type: string
        account_name:
          title: Account Name
          type: string
        id:
          format: uuid
          title: Organisation Id
          type: string
      required:
      - account_name
      - id
      title: Account
      type: object
...
    DeviceRead:
      properties:
        id:
          description: The account to which the device is registered, and into which
            GC Bundles will be issued for energy produced by this Device.
          format: uuid
          title: Account Id
          type: string
        capacity:
          title: Capacity
          type: number
        id:
          format: uuid
          title: Device Id
          type: string
        device_name:
          title: Device Name
          type: string
        energy_source:
          title: Energy Source
          type: string
        grid:
          title: Grid
          type: string
        location_id:
          format: uuid
          title: Location Id
          type: string
        operational_date:
          format: date
          title: Operational Date
          type: string
        peak_demand:
          title: Peak Demand
          type: number
        technology_type:
          title: Technology Type
          type: string
      required:
      - device_name
      - grid
      - energy_source
      - technology_type
      - operational_date
      - capacity
      - peak_demand
      - id
      - id
      title: DeviceRead
      type: object
    DeviceUpdate:
      properties:
        account:
          $ref: '#/components/schemas/Account'
        id:
          description: The account to which the device is registered, and into which
            GC Bundles will be issued for energy produced by this Device.
          format: uuid
          title: Account Id
          type: string
        capacity:
          title: Capacity
          type: number
        id:
          format: uuid
          title: Device Id
          type: string
        device_name:
          title: Device Name
          type: string
        energy_source:
          title: Energy Source
          type: string
        grid:
          title: Grid
          type: string
        location_id:
          format: uuid
          title: Location Id
          type: string
        meter_id:
          format: uuid
          title: Meter Id
          type: string
        operational_date:
          format: date
          title: Operational Date
          type: string
        peak_demand:
          title: Peak Demand
          type: number
        technology_type:
          title: Technology Type
          type: string
      required:
      - id
      title: DeviceUpdate
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    UserRead:
      properties:
        name:
          title: Name
          type: string
        id:
          format: uuid
          title: Organisation Id
          type: string
        email:
          title: Primary Contact
          type: string
        role:
          description: "The roles of the User with the registry. A single User can\
            \ be assigned multiple roles\n                       by the Registry Administrator\
            \ (which is itself a User for the purposes of managing allowable\n   \
            \                    actions), including: 'GC Issuer', 'Production Registrar',\
            \ 'Measurement Body', and 'Account Holder'.\n                       The\
            \ roles are used to determine the actions that the User is allowed to\
            \ perform within the registry,\n                       according to the\
            \ EnergyTag Standard."
          items:
            type: string
          title: Role
          type: array
        id:
          format: uuid
          title: User Id
          type: string
      required:
      - name
      - email
      - id
      - role
      - id
      title: UserRead
      type: object
    UserUpdate:
      properties:
        id:
          format: uuid
          title: Account Id
          type: string
        name:
          title: Name
          type: string
        id:
          format: uuid
          title: Organisation Id
          type: string
        email:
          title: Primary Contact
          type: string
        role:
          description: "The roles of the User with the registry. A single User can\
            \ be assigned multiple roles\n                       by the Registry Administrator\
            \ (which is itself a User for the purposes of managing allowable\n   \
            \                    actions), including: 'GC Issuer', 'Production Registrar',\
            \ 'Measurement Body', and 'Account Holder'.\n                       The\
            \ roles are used to determine the actions that the User is allowed to\
            \ perform within the registry,\n                       according to the\
            \ EnergyTag Standard."
          items:
            type: string
          title: Role
          type: array
        id:
          format: uuid
          title: User Id
          type: string
      required:
      - id
      - role
      title: UserUpdate
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: token
      type: oauth2
info:
  contact:
    email: connor@futureenergy.associates
    name: Please direct feedback to
  description: "\nThis document outlines the EnergyTag API Specification to accompany\
    \ Version 2 of \nthe EnergyTag Standard.\nThe functionality outlined within each\
    \ API interaction represents the recommended \nminimum required detail necessary\
    \ to implement a consistent Granular Certificate \n(GC) registry system. \nWithin\
    \ the Certificates section, the definition of GCs as specified in the Create \n\
    Certificate POST request, alongside the filtering parameters required to instigate\
    \ \nquery, transfer, and cancellation requests, will form core components of the\
    \ standard \nand impact the functionality of the registry the greatest. Subsequent\
    \ sections \ndescribing Organisation, User, Account, and Device management are\
    \ **recommendations \nonly** and do not form part of the standard, although terminology\
    \ used therein is \nreferenced in the certificate definitions. They are included\
    \ in the database class \ndiagram for the sake of completeness in representing\
    \ the relationships between the \ncertificate and action definitions; the fields\
    \ implemented for these tables are \nleft to the discretion of the registry operator.\
    \ \n<br>\n\nWhere Universally Unique Identifiers (UUIDs) have been used, these\
    \ can be replaced with \nany consistent and unique identification mechanism preferred\
...
openapi: 3.1.0
paths:
...
  /devices/device/{id}:
    delete:
      operationId: delete_device_devices_device__id__delete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Device Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Delete Device
      tags:
      - Devices
    get:
      operationId: read_device_devices_device__id__get
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Device Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Read Device
      tags:
      - Devices
    patch:
      operationId: update_device_devices_device__id__patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update Device
      tags:
      - Devices
...
  /users/user/{id}:
    delete:
      operationId: delete_user_users_user__id__delete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Delete User
      tags:
      - Users
    get:
      operationId: read_user_users_user__id__get
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Read User
      tags:
      - Users
    patch:
      operationId: update_user_users_user__id__patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update User
      tags:
      - Users
...